fix: long 타입 계산이 가능하도록 수정
- int 범위를 넘어서는 입력을 받을 수 있도록 수정
- 계산 결과로 long 타입을 반환하도록 수정

+++ src/main/java/calculator/io/output/ConsoleOutputHandler.java
@@ -10,7 +10,7 @@ public class ConsoleOutputHandler implements OutputHandler {
     }
 
     @Override
-    public void showAdditionResult(int calculateResult) {
+    public void showAdditionResult(long calculateResult) {
         System.out.println("결과 : " + calculateResult);
     }
