+++ src/main/java/calculator/operations/AdditionStrategy.java
@@ -5,9 +5,9 @@ import calculator.CalculatorException;
 public class AdditionStrategy implements CalculatorStrategy {
 
     @Override
-    public int calculate(int[] numbers) {
-        int result = 0;
-        for (int number : numbers) {
+    public long calculate(long[] numbers) {
+        long result = 0;
+        for (long number : numbers) {
             if (number < 0) {
                 throw new CalculatorException("양수만 입력할 수 있습니다.");
             }
